% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfclassic_general.R
\name{expand_stcvs}
\alias{expand_stcvs}
\title{Expand stanford classic variable strings}
\usage{
expand_stcvs(tibbl)
}
\arguments{
\item{tibbl}{a tibble having one row and any number of columns.
Each single 'tibbl[1,j]' is a string containing zero, one, or multiple variable
values, and where text variables are preceded by a newline
and where numeric variables are separated by a space}
}
\value{
a tibble, equal number of variables as input tibble.
}
\description{
This function takes a tibble having stanford classic variable values,
and expands these to single observations in a tibble (ala data frame)
All stanford variables in the input tibble should preferably
have equal number of observations.
}
\examples{
stanford_classic_variable_strings <-
tibble::tibble(speciescode = "1 2 3", speciesname = "\nspruce\npine\nbirch")
expand_stcvs (stanford_classic_variable_strings)
}
