% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfclassic_general.R
\name{file2strng}
\alias{file2strng}
\title{Read the content of a stanford classic text file and return the content as
a text string}
\usage{
file2strng(filename)
}
\arguments{
\item{filename}{Is the name of the stanford classic text file to read}
}
\value{
a string
}
\description{
Read the content of a stanford classic text file and return the content as
a text string
}
\examples{
tmp <- tempfile()
x <- "stanford classic is fun"
readr::write_file(x, tmp)
file2strng(tmp)
}
