% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drf2tables.R
\name{read_drf_file}
\alias{read_drf_file}
\title{Read StanForD Classic drf-files (machine reports from forest machines)}
\usage{
read_drf_file(filename)
}
\arguments{
\item{filename}{A file (including path) the function should read}
}
\value{
a list of tables populated with data from the drf report:
   report_header, object_definition, individual machine work time, combined machine work time
}
\description{
Read StanForD Classic drf-files (machine reports from forest machines)
}
\examples{
 files <- list.files(system.file("extdata", package = "stanfordclassicr"), full.names = TRUE)
 drffiles <- files[stringr::str_detect(files, ".drf")]
 drf_report <- read_drf_file(drffiles[1])
 drf_report <- read_drf_file(drffiles[4])
 drf_report <- read_drf_file(drffiles[5])
}
