% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktr2tables.R
\name{read_ktr_file}
\alias{read_ktr_file}
\title{Read StanForD Classic ktr-files (mahcine reports from forest machines)}
\usage{
read_ktr_file(filename)
}
\arguments{
\item{filename}{}
}
\value{

}
\description{
Read StanForD Classic ktr-files (mahcine reports from forest machines)
}
\examples{
 read_ktr_file(list.files(pattern = ".ktr")[1]

}
