% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prd2tables.R
\name{read_prd_file}
\alias{read_prd_file}
\title{Read StanForD Classic prd-files (machine reports from forest machines)}
\usage{
read_prd_file(filename)
}
\arguments{
\item{filename}{The file name (including path) of the file the
function should read}
}
\value{
should return a list of tables populated with data from the prd
report:
 report_header, object_definition, species_definition,
 product_definitions, volume_summary, logs_summary
}
\description{
Read StanForD Classic prd-files (machine reports from forest machines)
}
\examples{
 files <- list.files(system.file("extdata", package = "stanfordclassicr"), full.names = TRUE)
 prdfiles <- files[stringr::str_detect(files, ".prd")]
 prddata <- read_prd_file(prdfiles[1])
}
