% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pri2tables.R
\name{read_pri_file}
\alias{read_pri_file}
\title{Read StanForD Classic PRI-files (mahcine reports from forest machines)}
\usage{
read_pri_file(filename)
}
\arguments{
\item{filename}{}
}
\value{

}
\description{
Read StanForD Classic PRI-files (mahcine reports from forest machines)
}
\examples{
 read_pri_file(list.files(pattern = ".pri")[1])
}
