% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfclassic_general.R
\name{sfclassic2df}
\alias{sfclassic2df}
\title{Convert stanford classic text string to dataframe}
\usage{
sfclassic2df(strng)
}
\arguments{
\item{strng}{}
}
\value{
a data.frame, where each pair of StanForD code and corresponding
  variable values will form one column. I.e  have one row
}
\description{
Convert stanford classic text string to dataframe
}
\examples{
stanford_classic_string <- "~3 6 \n1470E~120 1 \nFURU \nGRAN \nLØV \nTØRRGRAN~"
sfclassic2df(stanford_classic_string)
}
