% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfvardefs_description.R
\docType{data}
\name{sfvardefs}
\alias{sfvardefs}
\title{StanForD classic variable definitions}
\format{An object of class tibble.}
\source{
\href{https://www.skogforsk.se/english/projects/stanford/}{StanForD documentation}
}
\usage{
data(sfvardefs)
}
\description{
This dataset provides all variable names types and classes (integer / character)
as defined in the StanForD classic Standard
}
\examples{
data(sfvardefs)
head(sfvardefs)
}
\references{
(\href{https://www.skogforsk.se/english/projects/stanford/}{whattowrite})
}
\keyword{datasets}
