% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktr2tables.R
\name{read_ktr_file}
\alias{read_ktr_file}
\title{Read StanForD Classic ktr-files (mahcine reports from forest machines)}
\usage{
read_ktr_file(filename)
}
\arguments{
\item{filename}{A filename (including path) the function should read}
}
\value{
a list of tables populated with data from the stm report: report_header, object_definition, calibration dates, control measurements
}
\description{
Read StanForD Classic ktr-files (mahcine reports from forest machines)
}
\examples{
 files <- list.files(system.file("extdata", package = "stanfordclassicr"), full.names = TRUE)
 ktrfiles <- files[stringr::str_detect(files, ".ktr")]
 filename = ktrfiles[2]
 ktrdata <- read_ktr_file(filename)

}
