% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pri2tables.R
\name{read_pri_file}
\alias{read_pri_file}
\title{Read StanForD Classic PRI-files (mahcine reports from forest machines)}
\usage{
read_pri_file(filename)
}
\arguments{
\item{filename}{}
}
\value{
should return a list of tables populated with data from the stm report: report_header, object_definition, operator_definition, product_definitions, stems, logs
}
\description{
Read StanForD Classic PRI-files (mahcine reports from forest machines)
}
\examples{
 files = list.files(system.file("extdata", package = "stanfordclassicr"), full.names = T)
 ktrfiles = files[stringr::str_detect(files, ".pri")]
 read_pri_file(prifiles[1])
}
