% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm2tables.R
\name{read_stm_file}
\alias{read_stm_file}
\title{Read StanForD Classic .stm-files (mahcine reports from forest machines)}
\usage{
read_stm_file(filename)
}
\arguments{
\item{filename}{}
}
\value{
a list of tables populated with data from the stm report: report_header, object_definition, operator_definition, product_definitions, stems, logs,
}
\description{
Read StanForD Classic .stm-files (mahcine reports from forest machines)
}
\details{
this function reads one stanford stemfile (.stm). Demo stanford stm files for testing the function:  setwd("./Demo StanFordClassic Reports/stm")
}
\examples{
read_stm_file(list.files(pattern = ".stm")[1])
}
