% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfordclassicr.R
\docType{package}
\name{stanfordclassicr}
\alias{stanfordclassicr}
\title{stanfordclassicr: A package for reading forest machine files following the
  stanford classic standard (https://www.skogforsk.se/english/projects/stanford/)}
\description{
The package provide one function for each type of stanford standard file
 and some supplementary functions for internal use.
}
